dev-3460: moved the etl code to a better home
