+++ usaspending_api/common/etl/__init__.py
@@ -1,0 +1,4 @@
+from usaspending_api.common.etl.etl_table import ETLTable
+
+
+__all__ = ["ETLTable"]
